+++ documentation/Changelog.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added dynamic fonts to ShoppingCartTableViewController and ShoppingCartTableCellView section #Apps-231
 * Added dynamic fonts to ShoppingCartViewController #Apps-231
 * Added dynamic fonts to CouponsListViewController #Apps-231
+* Added plural forms #APPS-352
 
 ### Removed
 * Xib/Nib utility methods